th-351: * close shift button to sidebar
