+++ frontend/src/pages/dashboard/components/sidebar/components/styles.module.scss
@@ -1,0 +1,15 @@
+@import "src/assets/css/vars.scss";
+
+.block {
+  margin-top: auto;
+}
+
+.button {
+  width: 100%;
+}
+
+.truckInfo {
+  margin: 10px 0;
+  color: $grey;
+  text-align: center;
+}
