+++ frontend/src/pages/orders/orders.tsx
@@ -1,22 +1,9 @@
-import { Button } from '~/libs/components/components.js';
 import { AppRoute } from '~/libs/enums/enums.js';
-import {
-  useAppDispatch,
-  useAppSelector,
-  useCallback,
-  useEffect,
-  useNavigate,
-} from '~/libs/hooks/hooks.js';
-import { actions as driverActions } from '~/slices/driver/driver.js';
+import { useAppSelector, useEffect, useNavigate } from '~/libs/hooks/hooks.js';
 import { ShiftStatus } from '~/slices/driver/libs/enums/enums.js';
-import {
-  selectActiveTruck,
-  selectShiftStatus,
-} from '~/slices/driver/selectors.js';
+import { selectShiftStatus } from '~/slices/driver/selectors.js';
 
 const Orders: React.FC = () => {
-  const dispatch = useAppDispatch();
-  const truck = useAppSelector(selectActiveTruck);
   const shiftStatus = useAppSelector(selectShiftStatus);
   const navigate = useNavigate();
   useEffect(() => {
@@ -24,18 +11,8 @@ const Orders: React.FC = () => {
       navigate(AppRoute.AVAILABLE_TRUCKS);
     }
   }, [shiftStatus, navigate]);
-  const handleClick = useCallback(() => {
-    void dispatch(driverActions.endShift());
-  }, [dispatch]);
 
-  return (
-    <div>
-      <p>
-        Your chosen truck: {truck?.manufacturer} ({truck?.id}) Test
-      </p>
-      <Button label={'End shift'} frontIcon={'truck'} onClick={handleClick} />
-    </div>
-  );
+  return <div>Orders page</div>;
 };
 
 export { Orders };
